requirements.txt
flask
sqlalchemy
psycopg2-binary
argon2-cffi